RTC_ID:1041046
COMMENT:[clea can]优化log输出，修改部分接口逻辑

PatchType:Task

PatchModule:NaviApp

PatchLevel:P2

+++ service/src/main/java/com/sgm/navi/service/adapter/signal/SignalAdapter.java
@@ -192,7 +192,7 @@ public final class SignalAdapter {
         mSignalApi.setDistanceToTrafficJamRoadAvailability(value);
     }
 
-    public void setDistanceOnTrafficJamRoad(int value) {
+    public void setDistanceOnTrafficJamRoad(float value) {
         if (mSignalApi == null) {
             return;
         }
